+++ src/deserialization.cpp
@@ -41,7 +41,7 @@ Status InputSink<ov::InferRequest&>::give(const std::string& name, ov::Tensor& t
 ov::Tensor makeTensor(const tensorflow::TensorProto& requestInput,
     const std::shared_ptr<TensorInfo>& tensorInfo) {
     ov::Shape shape;
-    for (size_t i = 0; i < static_cast<size_t>(requestInput.tensor_shape().dim_size()); i++) {
+    for (int i = 0; i < requestInput.tensor_shape().dim_size(); i++) {
         shape.push_back(requestInput.tensor_shape().dim(i).size());
     }
     ov::element::Type precision = tensorInfo->getOvPrecision();
@@ -52,7 +52,7 @@ ov::Tensor makeTensor(const ::inference::ModelInferRequest::InferInputTensor& re
     const std::shared_ptr<TensorInfo>& tensorInfo,
     const std::string& buffer) {
     ov::Shape shape;
-    for (size_t i = 0; i < static_cast<size_t>(requestInput.shape_size()); i++) {
+    for (int i = 0; i < requestInput.shape_size(); i++) {
         shape.push_back(requestInput.shape().at(i));
     }
     ov::element::Type precision = tensorInfo->getOvPrecision();
@@ -62,7 +62,7 @@ ov::Tensor makeTensor(const ::inference::ModelInferRequest::InferInputTensor& re
 ov::Tensor makeTensor(const ::inference::ModelInferRequest::InferInputTensor& requestInput,
     const std::shared_ptr<TensorInfo>& tensorInfo) {
     ov::Shape shape;
-    for (size_t i = 0; i < static_cast<size_t>(requestInput.shape_size()); i++) {
+    for (int i = 0; i < requestInput.shape_size(); i++) {
         shape.push_back(requestInput.shape().at(i));
     }
     ov::element::Type precision = tensorInfo->getOvPrecision();
