+++ src/tensorinfo.cpp
@@ -249,9 +249,9 @@ std::string TensorInfo::shapeToString(const shape_t& shape) {
 std::string tensorShapeToString(const google::protobuf::RepeatedField<int64_t>& shape) {
     std::ostringstream oss;
     oss << "(";
-    size_t i = 0;
+    int i = 0;
     if (shape.size() > 0) {
-        for (; i < static_cast<size_t>(shape.size()) - 1; i++) {
+        for (; i < shape.size() - 1; i++) {
             oss << shape[i] << ",";
         }
         oss << shape[i];
